Add files via upload
Project Name: Brawler Game Automation

Description:
The Brawler Game Automation project is a script developed to automate various tasks and actions in the Brawler blockchain game. The script interacts with the game's web interface using the Selenium WebDriver library and performs actions such as logging in, participating in brawls, collecting data, and healing the player's character when their health is low. The project aims to enhance the gameplay experience by automating repetitive tasks and providing notifications about important in-game events.

Key Features:

1.Login Automation: The script automates the login process to access the game's web interface.
2.Brawl Participation: It automates the process of participating in brawls within the game.
3.Data Collection: The script collects important game-related data such as token rates, gold amounts, and health status.
4.Healing Functionality: When the player's health falls below a specified threshold, the script automatically initiates the healing process to restore the character's health.
5.Line Integration: The project integrates with the Line messaging service to provide notifications about game events, including brawls, healing actions, and important game-related information.
Use Case:
The Brawler Game Automation project can be useful for players who want to streamline their gameplay experience in the Brawler blockchain game. It saves time and effort by automating repetitive tasks, allowing players to focus on other aspects of the game. The integration with the Line messaging service enables players to receive real-time notifications about in-game events and important updates, even when they are not actively playing the game. The project enhances the overall efficiency and convenience of playing the Brawler blockchain game.